% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_CNAE.R
\docType{data}
\name{cl_CNAE}
\alias{cl_CNAE}
\title{Classification labels of the CNAE2 data set}
\format{
A vector containing the true classification labels of documents of \code{CNAE2} data set.
}
\source{
See \code{CNAE2} for further details.
}
\usage{
data(CNAE2)
}
\description{
True classification labels of documents from the \code{CNAE2} data set.
}
\keyword{datasets}
